feat: Add `JSContext::global_object`.
This patch adds the `global_object` method on the `JSContext`
struct. The idea is to be able to retrive the global object. It uses
`JSContextGetGlobalObject` behind the scene.

Because a global object might not always be available, `global_object`
returns an `Result<JSObject, JSException>`.

This patch adds a basic doctest, and a simple unit test.