fix: `#[function_callback]` can be used without `javascriptcore_sys`.
This patch fixes the `#[function_callback]` procedural macro. It
was using types from `javascriptcore` and from `javascriptcore_sys`.
Problem: if the user uses `javascriptcore` only, and wants to use the
macro, it has to add `javascriptcore_sys` as a dependency.

To avoid this situation, we expose `javascript_sys` as a (doc hidden)
module from `javascript` under the `sys` name.

This patch also updates the `#[function_callback]` to use
`javascriptcore::sys` instead of `javascriptcore`, and we are good.