feat: Add a `JSClassBuilder`.
This patch revamps `JSClass` by introducing `JSClassBuilder`. A
`JSClass` is kind of rich, and providing a clean API is difficult, hence
the `JSClassBuilder` which brings a nice API for building any kind of
`JSClass`. For now, only the classname and the constructor callback can
be set, but this pattern is easily extendable.

This patch also adds documentation and improves the tests a little bit.