Release JSPropertyNameArrayRef. Retain string.
Since `JSObjectCopyPropertyNames` returns the name array under the
create rule, we need to release it when it is dropped.

Since it owns the strings, and the strings can outlive the name
array (like when iterating over the names and storing the string
references), we need to retain them to keep them alive.